Fix: style conflict with multiple gauges
When multiple gauges are created utilizing the "parentNode" method, the obj.config.id is undefined. When the "generateShadow" function is called, the filter is created as "inner-shadow-" for each gauge. This creates a conflict between each gauge on the screen.

Attempting to set an "id" didn't work because this took priority and the code looked for the "id" to exist in the dom instead of using the parentNode.

I created the classId property to set any arbitrary id to be used as the filter id. This will prevent conflicts and can be used when the parentNode method is used.

If a classId is not set, it will default to the previous mechanism.